Refactor: Add ProjectWithImages interface
Added the ProjectWithImages interface to the type.ts file in the lib/schemas directory. This interface extends the Project interface and includes an array of Image objects. This change is necessary for handling projects with associated images.

Refactor: Remove unused code in +page.svelte

Removed unused code in the +page.svelte file in the src/routes/[[lang=lang]]/(app)/ directory. The removed code includes the import statement for the PageData type and the createCounter function. This cleanup improves code readability and reduces unnecessary complexity.

Refactor: Update Image component in Image.svelte

Updated the Image component in the Image.svelte file in the lib/components/image/ directory. The change involves replacing the hard-coded src attribute value with a dynamic value passed as a prop. This allows for more flexibility in displaying different images based on the prop value.

Refactor: Update Prisma client import in prisma.ts

Updated the import statement for the Prisma client in the prisma.ts file in the lib/server/ directory. The change involves commenting out the import statement for the '@prisma/client/edge' package and uncommenting the import statement for the '@prisma/client' package. This change is necessary to make the code compatible with the local development environment.

Refactor: Update projects/+page.ts and projects/+server.ts

Updated the projects/+page.ts and projects/+server.ts files in the src/routes/[[lang=lang]]/(app)/projects/ and src/routes/api/projects/ directories, respectively. The changes involve removing commented-out code related to prerendering and updating the cacheStrategy options in the GET request handler. These changes improve code readability and ensure proper caching behavior.

Refactor: Update schema.prisma

Updated the schema.prisma file in the prisma/ directory. The changes involve adding the 'index' and 'highlighted' fields to the Project model, and the 'highlighted' field to the Image and Video models. These changes allow for better organization and filtering of projects, images, and videos based on their index and highlighted status.